requirements.txt
# prompt-generation
langchain>=0.2.0
langchain-openai
langchain-community
boto3>=1.34.0
botocore>=1.34.0
python-dotenv>=1.0.0
chromadb

# chromadb ui + streamlit dashboard
fastapi
uvicorn
matplotlib
scikit-learn
pandas
plotly
altair
streamlit
umap-learn

# activation capture
transformer_lens
sae-lens>=6.5.3
tqdm
numpy
wandb
datasets

# infrastructure, runpod
ruff
nvitop
httpx[socks]
sagemaker